@@ -16,19 +16,19 @@
         }
     ],
     "require": {
-        "php": ">=7.2.5",
-        "symfony/polyfill-mbstring": "~1.0",
-        "symfony/polyfill-php80": "^1.16"
+        "php": ">=8.0.2",
+        "symfony/polyfill-mbstring": "~1.0"
     },
     "require-dev": {
         "ext-iconv": "*",
-        "symfony/console": "^4.4|^5.0",
-        "symfony/process": "^4.4|^5.0",
+        "symfony/console": "^5.4|^6.0",
+        "symfony/process": "^5.4|^6.0",
+        "symfony/uid": "^5.4|^6.0",
         "twig/twig": "^2.13|^3.0.4"
     },
     "conflict": {
         "phpunit/phpunit": "<5.4.3",
-        "symfony/console": "<4.4"
+        "symfony/console": "<5.4"
     },
     "suggest": {
         "ext-iconv": "To convert non-UTF-8 strings to UTF-8 (or symfony/polyfill-iconv in case ext-iconv cannot be used).",
